requirements.txt
python-telegram-bot[all]==20.3
gspread==5.10.0
oauth2client==4.1.3
pandas==2.0.3
asyncpg~=0.27.0
aiohttp~=3.8.4
SQLAlchemy~=2.0.17
numpy~=1.25.0